domain specific abstraction
