+++ src/components/Header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  PageHeader, Space, Button,
+  PageHeader, Space,
 } from 'antd';
 import PropTypes from 'prop-types';
 import { NextSeo } from 'next-seo';
@@ -10,7 +10,8 @@ import UserButton from 'components/header/UserButton';
 import FeedbackButton from 'components/header/FeedbackButton';
 import ReferralButton from 'components/header/ReferralButton';
 import HelpButton from 'components/header/HelpButton';
-import { DomainName } from 'utils/deploymentInfo';
+import nextConfig from 'next/config';
+import returnDomainSpecificContent from 'utils/domainSpecificContent.jsx';
 
 const Header = (props) => {
   const {
@@ -18,10 +19,10 @@ const Header = (props) => {
   } = props;
   const experiment = useSelector((state) => state?.experiments[experimentId]);
   const experimentName = experimentData?.experimentName || experiment?.name;
-  const domainName = useSelector((state) => state.networkResources?.domainName);
   const truncateText = (text) => (
     (text && text.length > 28) ? `${text.substr(0, 27)}…` : text
   );
+  const accountId = nextConfig()?.publicRuntimeConfig?.accountId;
 
   return (
     <>
@@ -36,15 +37,10 @@ const Header = (props) => {
         extra={(
           <Space size='large'>
             <Space>
-              {(domainName === DomainName.BIOMAGE || domainName === DomainName.BIOMAGE_STAGING)
-                && (
-                  <Button>
-                    <a href='https://www.biomage.net/cellenicscourse' target='_blank' rel='noreferrer'>
-                      Courses
-                    </a>
-                  </Button>
-                )}
-              <HelpButton />
+              {returnDomainSpecificContent('Courses', accountId)}
+              <HelpButton
+                accountId={accountId}
+              />
               <FeedbackButton />
               <ReferralButton />
               {extra}
